Major refactoring of code structure, add new testframework
Since long time run of memshare code made each process
attach to shared memory time after time.
That could eventually result in lack of resources.

To verify that this is not happening a new test framework
is added. This framework spawnes 19 processes and keeps
sending (at random) signals to these processes.
Equally random a process is terminated and started again.